created a separate Router's class for a client-side of application
